requirements.txt
python-dotenv
flask
google-generativeai
requests
streamlit
beautifulsoup4
langchain
flask_swagger_ui

